bot/requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
pydantic==2.5.2
pydantic_core==2.14.5
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
aio-pika==9.3.1
pika==1.3.2
psycopg2-binary
